requirements.txt
flask==3.0.2
flask-sqlalchemy
python-dotenv
